Extract reportError from util.ts to be consistent with getColorScale and createColormap
