+++ src/util.ts
@@ -1,6 +1,4 @@
 import * as config from './clientConfig.json';
-import { ElNotification } from 'element-ui/types/notification'
-import * as Raven from 'raven-js';
 
 const fuConfig = config.fineUploader;
 
@@ -77,22 +75,6 @@ function mdTypeSupportsOpticalImages(mdType: string): boolean {
   return !mdTypesToSkipImages.includes(mdType);
 }
 
-let $notify: ElNotification;
-function setErrorNotifier(_$notify: ElNotification) {
-  $notify = _$notify;
-}
-
-function reportError(err: Error, message?: string) {
-  try {
-    Raven.captureException(err);
-    if ($notify != null) {
-      $notify.error(message || 'Oops! Something went wrong. Please refresh the page and try again.');
-    }
-  } catch(ex) {
-    console.error(ex);
-    /* Avoid breaking down-stream error handling  */
-  }
-}
 
 export {
   renderMolFormula,
@@ -105,6 +87,4 @@ export {
   csvExportHeader,
   scrollDistance,
   mdTypeSupportsOpticalImages,
-  setErrorNotifier,
-  reportError,
 };
